requirements.txt
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.41.0
bleach==6.1.0
blinker==1.7.0
cachetools==5.5.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.27.7
cog==0.8.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
datasets==4.0.0
debugpy==1.8.1
decorator==5.1.1
deepspeed==0.9.5
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.6.1
einops-exts==0.0.4
et_xmlfile==2.0.0
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.98.0
fastjsonschema==2.19.1
ffmpy==0.3.1
filelock==3.13.1
Flask==3.0.3
fonttools==4.44.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
h11==0.14.0
hjson==3.1.0
httpcore==0.17.3
httplib2==0.22.0
httptools==0.6.1
huggingface==0.0.1
huggingface-hub==0.25.1
idna==3.4
importlib-metadata==7.0.1
ipykernel==6.29.3
ipython==8.12.3
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
json5==0.9.20
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
line-profiler==4.1.2
linkify-it-py==2.0.2
lit==17.0.4
# llama_cpp_python==0.2.90  # Installed separately with system tools
git+https://github.com/haotian-liu/LLaVA.git@786aa6a19ea10edc6f574ad2e16276974e9aaa3a#egg=llava
markdown-it-py==2.2.0
markdown2==2.4.10
MarkupSafe==3.0.0
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.16
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
openpyxl==3.1.5
orjson==3.9.10
overrides==7.7.0
packaging==23.2
pandas==2.1.2
pandocfilters==1.5.1
parso==0.8.3
peft==0.4.0
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
pipreqs==0.5.0
platformdirs==4.1.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.7
pycparser==2.21
pydantic==1.10.13
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.1.1
pytesseract==0.3.13
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.30.2
regex==2023.10.3
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.12.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.2.2
scipy==1.11.3
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.34.0
setproctitle==1.3.3
shortuuid==1.0.11
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.27.0
structlog==23.2.0
supervision==0.17.1
svgwrite==1.4.3
sympy==1.12
terminado==0.18.0
threadpoolctl==3.2.0
timm==0.6.13
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch>=2.1.0
torchvision>=0.16.0
tornado==6.4
tqdm==4.67.1
traitlets==5.14.1
transformers>=4.40.0
triton==2.0.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.12.2
tzdata==2023.3
uc-micro-py==1.0.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.24.0.post1
uvloop==0.19.0
wandb==0.16.0
watchfiles==0.21.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.2
xxhash==3.5.0
yapf==0.40.2
yarg==0.1.9
yarl==1.9.2
zipp==3.17.0
